Moved ID into the BaseEntity class and added the addID method to the AbstractService
