+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/BaseEntity.java
@@ -1,5 +1,20 @@
 package uk.gov.companieshouse.api.accounts.model.entity;
 
-public interface BaseEntity {
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Field;
+
+public class BaseEntity {
+
+    @Id
+    @Field("_id")
+    private String id;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
 
 }
