+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/CompanyAccountEntity.java
@@ -1,26 +1,13 @@
 package uk.gov.companieshouse.api.accounts.model.entity;
 
-import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
 @Document(collection = "accounts")
-public class CompanyAccountEntity implements BaseEntity{
-
-    @Id
-    @Field("_id")
-    private String id;
+public class CompanyAccountEntity extends BaseEntity{
 
     @Field("data")
     private CompanyAccountDataEntity data;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public CompanyAccountDataEntity getData() {
         return data;
