+++ src/main/java/uk/gov/companieshouse/api/accounts/service/AbstractService.java
@@ -1,6 +1,5 @@
 package uk.gov.companieshouse.api.accounts.service;
 
-import java.util.Map;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.api.accounts.model.entity.BaseEntity;
 import uk.gov.companieshouse.api.accounts.model.rest.RestObject;
@@ -15,5 +14,7 @@ public interface AbstractService<C extends RestObject, E extends BaseEntity> {
     public void addLinks(C rest);
 
     public void addKind(C rest);
+
+    public void addID(E entity);
 
 }
